Compute links eagerly
So that they can be included in JSON metadata
